# backend/src/spec/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"DeckCard": {
				"properties": {
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"cardId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"quantity",
					"cardId"
				],
				"type": "object"
			},
			"DeckDTO": {
				"properties": {
					"versionId": {
						"type": "string"
					},
					"attributes": {
						"items": {
							"properties": {
								"value": {
									"type": "string"
								},
								"name": {
									"type": "string"
								}
							},
							"required": [
								"value",
								"name"
							],
							"type": "object"
						},
						"type": "array"
					},
					"deckId": {
						"type": "string"
					},
					"mana": {
						"type": "string"
					},
					"mainDeck": {
						"items": {
							"$ref": "#/components/schemas/DeckCard"
						},
						"type": "array"
					},
					"reducedSideboard": {
						"items": {
							"$ref": "#/components/schemas/DeckCard"
						},
						"type": "array"
					},
					"sideboard": {
						"items": {
							"$ref": "#/components/schemas/DeckCard"
						},
						"type": "array"
					},
					"commandZone": {},
					"companions": {
						"items": {
							"$ref": "#/components/schemas/DeckCard"
						},
						"type": "array"
					},
					"hash": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"active": {
						"type": "boolean"
					},
					"winrate": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"versionId",
					"attributes",
					"deckId",
					"mana",
					"mainDeck",
					"reducedSideboard",
					"sideboard",
					"commandZone",
					"companions",
					"hash",
					"name",
					"active",
					"winrate",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGameResult": {
				"properties": {
					"playerDeckId": {
						"type": "string"
					},
					"opponentDeckColors": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"result": {
						"type": "string"
					},
					"oponent": {
						"type": "string"
					},
					"matchId": {
						"type": "string"
					},
					"onThePlay": {
						"type": "boolean"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"playerDeckId",
					"opponentDeckColors",
					"result",
					"oponent",
					"matchId",
					"onThePlay",
					"date"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "mrg-tracker",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Pau Sala"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/decks": {
			"get": {
				"operationId": "GetDecks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/DeckDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/games/version/{deckId}": {
			"get": {
				"operationId": "GetDecksByVersion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IGameResult"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deckId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/games/{deckId}": {
			"get": {
				"operationId": "GetDecks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IGameResult"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deckId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}